+++ lib/collab-comms/src/index.ts
@@ -1,15 +1,21 @@
 export { ClientCommunication } from './client-communication';
+export type { Message } from './collab-message-bus';
 export { CollabMessageBus, MessageType } from './collab-message-bus';
 export type {
-  CollabRequest,
-  CollabRequestGetDocument,
-  CollabRequestPullEvents,
-  CollabRequestPushEvents,
+  CollabClientRequest,
+  CollabClientRequestGetDocument,
+  CollabClientRequestPullEvents,
+  CollabClientRequestPushEvents,
   GetDocumentResponseBody,
   PullEventsRequestBody,
   PullEventsResponseBody,
   PushEventsRequestBody,
   PushEventsResponseBody,
 } from './common';
-export { CollabFail, CollabRequestType, NetworkingError } from './common';
+export {
+  CollabClientRequestType,
+  CollabFail,
+  CollabManagerBroadCastType,
+  NetworkingError,
+} from './common';
 export { ManagerCommunication } from './manager-communication';